server/requirements.txt
weaviate-client==4.7.1